Pytest plugins in their own file
